int vs float .div

+++ programs/meme-race/src/ix/claim_with_wager.rs
@@ -23,7 +23,7 @@ pub fn ix(ctx: Context<ClaimWithWager>) -> Result<()> {
         // compute winners pot
         let pot = (leader_board.total as f32) * pda::boss::POT_SPLIT;
         // floor div bc this dog coin has got too many digits anyways
-        let share_pct = wager.wager_size.div(winner.score) as f32;
+        let share_pct = (wager.wager_size as f32).div(winner.score as f32);
         let share = (pot * share_pct) as u64;
         // build transfer ix
         let transfer_accounts = Transfer {
